[ENG-7461] Fix Validation Issues with Article DOIs (#11062)
## Purpose
Article DOI aren't validated properly, this is going to address the problem and fix it.

## Changes
- Splitting up the one big preprint detail test file into five file ~< 500 lines, so I can zero in on the correct behavior.
- Adds new DOIField to `doi` on preprint serializer.
- remove superfluous preprint save statements for preprint serializer
- Moved article DOI validation into the Preprint model `full_clean` to prevent future issues.

## Ticket
https://openscience.atlassian.net/browse/ENG-7461
