+++ api_tests/preprints/views/test_preprint_detail.py
@@ -1,28 +1,9 @@
-from unittest import mock
 import pytest
-import datetime
-import responses
 
 from django.utils import timezone
-from rest_framework import exceptions
-from waffle.testutils import (
-    override_switch,
-)
 
-from osf import features
-from osf.utils.permissions import READ
 from api.base.settings.defaults import API_BASE
-from api_tests import utils as test_utils
-from api_tests.subjects.mixins import UpdateSubjectsMixin
-from framework.auth.core import Auth
-from osf.models import (
-    NodeLicense,
-    PreprintContributor,
-    PreprintLog
-)
-from osf.utils import permissions as osf_permissions
 from osf.utils.permissions import WRITE, ADMIN
-from osf.utils.workflows import DefaultStates
 from osf_tests.factories import (
     PreprintFactory,
     AuthUserFactory,
@@ -31,7 +12,7 @@ from osf_tests.factories import (
     PreprintProviderFactory,
     InstitutionFactory
 )
-from website.settings import DOI_FORMAT, CROSSREF_URL
+from website.settings import DOI_FORMAT
 
 
 def build_preprint_update_payload(
@@ -270,1615 +251,6 @@ class TestPreprintDelete:
 
 
 @pytest.mark.django_db
-@pytest.mark.enable_enqueue_task
-class TestPreprintUpdate:
-
-    @pytest.fixture()
-    def preprint(self, user):
-        return PreprintFactory(creator=user, provider=PreprintProviderFactory(doi_prefix='10.1234/test'))
-
-    @pytest.fixture()
-    def url(self, preprint):
-        return f'/{API_BASE}preprints/{preprint._id}/'
-
-    @pytest.fixture()
-    def subject(self):
-        return SubjectFactory()
-
-    def test_update_preprint_permission_denied(self, app, preprint, url):
-        update_doi_payload = build_preprint_update_payload(
-            preprint._id, attributes={'article_doi': '10.123/456/789'})
-
-        noncontrib = AuthUserFactory()
-        res = app.patch_json_api(
-            url,
-            update_doi_payload,
-            auth=noncontrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-
-        res = app.patch_json_api(url, update_doi_payload, expect_errors=True)
-        assert res.status_code == 401
-
-        read_contrib = AuthUserFactory()
-        preprint.add_contributor(read_contrib, READ, save=True)
-        res = app.patch_json_api(
-            url,
-            update_doi_payload,
-            auth=read_contrib.auth,
-            expect_errors=True
-        )
-        assert res.status_code == 403
-
-    def test_update_original_publication_date_to_none(self, app, preprint, url):
-        # Original pub date accidentally set, need to remove
-        write_contrib = AuthUserFactory()
-        preprint.add_contributor(write_contrib, WRITE, save=True)
-        preprint.original_publication_date = '2013-12-11 10:09:08.070605+00:00'
-        preprint.save()
-        update_payload = build_preprint_update_payload(
-            preprint._id, attributes={
-                'original_publication_date': None,
-            }
-        )
-
-        res = app.patch_json_api(
-            url,
-            update_payload,
-            auth=write_contrib.auth,
-        )
-
-        assert res.status_code == 200
-        preprint.reload()
-        assert preprint.original_publication_date is None
-
-    def test_update_custom_publication_citation_to_none(self, app, preprint, url):
-        write_contrib = AuthUserFactory()
-        preprint.add_contributor(write_contrib, WRITE, save=True)
-        preprint.custom_publication_citation = 'fake citation'
-        preprint.save()
-        update_payload = build_preprint_update_payload(
-            preprint._id, attributes={'custom_publication_citation': None}
-        )
-        res = app.patch_json_api(url, update_payload, auth=write_contrib.auth)
-        assert res.status_code == 200
-        preprint.reload()
-        assert preprint.custom_publication_citation is None
-
-    @responses.activate
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated', mock.Mock())
-    def test_update_preprint_permission_write_contrib(self, app, preprint, url):
-        responses.add(
-            responses.Response(
-                responses.POST,
-                CROSSREF_URL,
-                content_type='text/html;charset=ISO-8859-1',
-                status=200,
-            ),
-        )
-        write_contrib = AuthUserFactory()
-        preprint.add_contributor(write_contrib, WRITE, save=True)
-
-        doi = '10.123/456/789'
-        original_publication_date = '2013-12-11 10:09:08.070605+00:00'
-        license_record = {
-            'year': '2015',
-            'copyright_holders': ['Tonya Bateman']
-        }
-        license = NodeLicense.objects.filter(name='No license').first()
-        title = 'My Preprint Title'
-        description = 'My Preprint Description'
-        tags = ['test tag']
-        node = ProjectFactory(creator=write_contrib)
-        new_file = test_utils.create_test_preprint_file(
-            preprint, write_contrib, filename='shook_that_mans_hand.pdf')
-
-        update_payload = build_preprint_update_payload(
-            preprint._id, attributes={
-                'original_publication_date': original_publication_date,
-                'doi': doi,
-                'license_record': license_record,
-                'title': title,
-                'description': description,
-                'tags': tags,
-            }, relationships={'node': {'data': {'type': 'nodes', 'id': node._id}},
-            'primary_file': {'data': {'type': 'file', 'id': new_file._id}},
-                'license': {'data': {'type': 'licenses', 'id': license._id}}}
-        )
-
-        res = app.patch_json_api(
-            url,
-            update_payload,
-            auth=write_contrib.auth,
-        )
-
-        assert res.status_code == 200
-        preprint.reload()
-
-        assert preprint.article_doi == doi
-        assert str(preprint.original_publication_date) == original_publication_date
-        assert preprint.license.node_license == license
-        assert preprint.license.year == license_record['year']
-        assert preprint.license.copyright_holders == license_record['copyright_holders']
-        assert preprint.title == title
-        assert preprint.description == description
-        assert preprint.tags.first().name == tags[0]
-        assert preprint.node == node
-        assert preprint.primary_file == new_file
-
-    def test_update_published_write_contrib(self, app, preprint, url):
-        preprint.is_published = False
-        preprint.save()
-
-        write_contrib = AuthUserFactory()
-        preprint.add_contributor(write_contrib, WRITE, save=True)
-
-        update_payload = build_preprint_update_payload(
-            preprint._id, attributes={
-                'is_published': 'true'
-            }
-        )
-
-        res = app.patch_json_api(
-            url,
-            update_payload,
-            auth=write_contrib.auth,
-            expect_errors=True)
-
-        assert res.status_code == 403
-        assert preprint.is_published is False
-
-    def test_update_node(self, app, user, preprint, url):
-        assert preprint.node is None
-        node = ProjectFactory(creator=user)
-        update_node_payload = build_preprint_update_payload(
-            preprint._id, relationships={'node': {'data': {'type': 'nodes', 'id': node._id}}}
-        )
-
-        res = app.patch_json_api(url, update_node_payload, auth=user.auth)
-        assert res.status_code == 200
-        assert res.json['data']['relationships']['node']['data']['id'] == node._id
-        preprint.reload()
-        assert preprint.node == node
-
-    def test_update_node_permissions(self, app, user, preprint, url):
-        assert preprint.node is None
-        node = ProjectFactory()
-        update_node_payload = build_preprint_update_payload(
-            preprint._id, relationships={'node': {'data': {'type': 'nodes', 'id': node._id}}}
-        )
-
-        res = app.patch_json_api(url, update_node_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 403
-        preprint.reload()
-        assert preprint.node is None
-
-    def test_update_node_existing_preprint(self, app, user, preprint, url):
-        assert preprint.node is None
-        node = ProjectFactory(creator=user)
-        # Create preprint with same provider on node
-        PreprintFactory(creator=user, project=node, provider=preprint.provider)
-        update_node_payload = build_preprint_update_payload(
-            preprint._id, relationships={'node': {'data': {'type': 'nodes', 'id': node._id}}}
-        )
-
-        res = app.patch_json_api(url, update_node_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 200
-        preprint.reload()
-        assert preprint.node == node
-
-    def test_update_deleted_node(self, app, user, preprint, url):
-        assert preprint.node is None
-        node = ProjectFactory(creator=user)
-        node.is_deleted = True
-        node.save()
-        update_node_payload = build_preprint_update_payload(
-            preprint._id, relationships={'node': {'data': {'type': 'nodes', 'id': node._id}}}
-        )
-
-        res = app.patch_json_api(url, update_node_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Cannot attach a deleted project to a preprint.'
-        preprint.reload()
-        assert preprint.node is None
-
-    def test_update_primary_file(self, app, user, preprint, url):
-        new_file = test_utils.create_test_preprint_file(
-            preprint, user, filename='shook_that_mans_hand.pdf')
-        relationships = {
-            'primary_file': {
-                'data': {
-                    'type': 'file',
-                    'id': new_file._id
-                }
-            }
-        }
-        assert preprint.primary_file != new_file
-        update_file_payload = build_preprint_update_payload(
-            preprint._id, relationships=relationships)
-
-        res = app.patch_json_api(url, update_file_payload, auth=user.auth)
-        assert res.status_code == 200
-
-        preprint.reload()
-        assert preprint.primary_file == new_file
-
-        log = preprint.logs.latest()
-        assert log.action == 'file_updated'
-        assert log.params.get('preprint') == preprint._id
-
-    def test_update_preprints_with_none_type(self, app, user, preprint, url):
-        payload = {
-            'data': {
-                'id': preprint._id,
-                'type': None,
-                'attributes': None,
-                'relationship': None
-            }
-        }
-
-        res = app.patch_json_api(url, payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['source']['pointer'] == '/data/type'
-
-    def test_update_preprints_with_no_type(self, app, user, preprint, url):
-        payload = {
-            'data': {
-                'id': preprint._id,
-                'attributes': None,
-                'relationship': None
-            }
-        }
-
-        res = app.patch_json_api(url, payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['source']['pointer'] == '/data/type'
-
-    def test_update_preprints_with_wrong_type(self, app, user, preprint, url):
-        update_file_payload = build_preprint_update_payload(preprint._id, jsonapi_type='Nonsense')
-
-        res = app.patch_json_api(url, update_file_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 409
-
-    def test_new_primary_not_in_node(self, app, user, preprint, url):
-        project = ProjectFactory()
-        file_for_project = test_utils.create_test_file(
-            project, user, filename='six_pack_novak.pdf')
-
-        relationships = {
-            'primary_file': {
-                'data': {
-                    'type': 'file',
-                    'id': file_for_project._id
-                }
-            }
-        }
-
-        update_file_payload = build_preprint_update_payload(
-            preprint._id, relationships=relationships)
-
-        res = app.patch_json_api(
-            url, update_file_payload,
-            auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-
-        preprint.reload()
-        assert preprint.primary_file != file_for_project
-
-    def test_update_original_publication_date(self, app, user, preprint, url):
-        date = timezone.now() - datetime.timedelta(days=365)
-        update_payload = build_preprint_update_payload(
-            preprint._id, attributes={'original_publication_date': str(date)}
-        )
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-        assert res.status_code == 200
-
-        preprint.reload()
-        assert preprint.original_publication_date == date
-
-    def test_update_custom_publication_citation(self, app, user, preprint, url):
-        citation = 'fake citation'
-        update_payload = build_preprint_update_payload(
-            preprint._id, attributes={'custom_publication_citation': citation}
-        )
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-        assert res.status_code == 200
-        preprint.reload()
-        assert preprint.custom_publication_citation == citation
-
-    @responses.activate
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated', mock.Mock())
-    def test_update_article_doi(self, app, user, preprint, url):
-        responses.add(
-            responses.Response(
-                responses.POST,
-                CROSSREF_URL,
-                content_type='text/html;charset=ISO-8859-1',
-                status=200,
-            ),
-        )
-
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'doi': '10.1235/test'}
-        )
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-        assert res.status_code == 200
-
-        preprint_doi = DOI_FORMAT.format(prefix=preprint.provider.doi_prefix, guid=preprint._id)
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'doi': preprint_doi}
-        )
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-        error_data = res.json['errors']
-        assert ' is already associated with this preprint' in error_data[0]['detail']
-
-        preprint.reload()
-        assert preprint.article_doi == '10.1235/test'
-
-        preprint_detail = app.get(url, auth=user.auth).json['data']
-        assert preprint_detail['links']['doi'] == 'https://doi.org/10.1235/test'
-
-    @responses.activate
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated', mock.Mock())
-    def test_update_article_doi_different_formats(self, app, user, preprint, url):
-        responses.add(
-            responses.Response(
-                responses.POST,
-                CROSSREF_URL,
-                content_type='text/html;charset=ISO-8859-1',
-                status=200,
-            ),
-        )
-        dois = [
-            ('', False),
-            (None, False),
-            ('10.1235/', False),
-            ('10.1235/test', False),
-            ('doi.org/10.1235/test', False),
-            ('https://10.1235/test', False),
-            ('https://doi.org/10.1235/', False),
-            ('https://doi.org/10.1235/test', False),
-            ('https://doi.org/10.1235/osf.io/12345', False),
-            ('https://doi.org/10.1235/test-1234-01-02', False),
-            ('10.1235', True),
-            ('9.1234/test', True),
-            ('https://doi.org/9.1234/test', True),
-            ('https://doi.org/9.1234/', True),
-            ('https://9.1234/test', True),
-            ('doi.org/9.1234/test', True),
-            ('9.1234', True),
-        ]
-        for doi, expect_error in dois:
-            update_payload = build_preprint_update_payload(
-                preprint._id,
-                attributes={'doi': doi}
-            )
-            res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=expect_error)
-            assert res.status_code == 400 if expect_error else 200
-
-    @responses.activate
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated', mock.Mock())
-    def test_article_doi_regex(self, app, user, preprint, url):
-        responses.add(
-            responses.Response(
-                responses.POST,
-                CROSSREF_URL,
-                content_type='text/html;charset=ISO-8859-1',
-                status=200,
-            ),
-        )
-        original_and_stripped_dois = [
-            ('10.1235/', '10.1235/'),
-            ('10.1235/test', '10.1235/test'),
-            ('doi.org/10.1235/test', '10.1235/test'),
-            ('https://10.1235/test', '10.1235/test'),
-            ('https://doi.org/10.1235/', '10.1235/'),
-            ('https://doi.org/10.1235/test', '10.1235/test'),
-            ('10.1312321321/test/tes2/test/13', '10.1312321321/test/tes2/test/13'),
-            ('10.1235/word1/random/preprint/id/any/other/string///doi', '10.1235/word1/random/preprint/id/any/other/string///doi'),
-            ('http://doi.com/10.1236/test/url/http', '10.1236/test/url/http'),
-            ('10.123/456/789', '10.123/456/789'),
-            ('10.word/', '10.word/'),
-            ('word.10.1234/word2', '10.1234/word2'),
-            ('10.word/word2/', '10.word/word2/'),
-            ('https://doi.org/10.1235/osf.io/12345', '10.1235/osf.io/12345'),
-            ('https://doi.org/10.1235/test-1234-01-02', '10.1235/test-1234-01-02'),
-        ]
-        for doi, stripped_doi in original_and_stripped_dois:
-            update_payload = build_preprint_update_payload(
-                preprint._id,
-                attributes={'doi': doi}
-            )
-            app.patch_json_api(url, update_payload, auth=user.auth)
-
-            preprint_detail = app.get(url, auth=user.auth).json['data']
-            assert preprint_detail['links']['doi'] == f'https://doi.org/{stripped_doi}'
-
-    def test_title_has_a_512_char_limit(self, app, user, preprint, url):
-        new_title = 'a' * 513
-        update_title_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'title': new_title,
-            }
-        )
-        res = app.patch_json_api(
-            url,
-            update_title_payload,
-            auth=user.auth,
-            expect_errors=True
-        )
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Ensure this field has no more than 512 characters.'
-        preprint.reload()
-        assert preprint.title != new_title
-
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated')
-    def test_update_description_and_title(
-            self, mock_preprint_updated, app, user, preprint, url):
-        new_title = 'Brother Nero'
-        new_description = 'I knew you\'d come!'
-        assert preprint.description != new_description
-        assert preprint.title != new_title
-        update_title_description_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'title': new_title,
-                'description': new_description,
-            }
-        )
-        res = app.patch_json_api(
-            url,
-            update_title_description_payload,
-            auth=user.auth)
-
-        assert res.status_code == 200
-        preprint.reload()
-
-        assert preprint.description == new_description
-        assert preprint.title == new_title
-        assert mock_preprint_updated.called
-
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated')
-    def test_update_tags(self, mock_update_doi_metadata, app, user, preprint, url):
-        new_tags = ['hey', 'sup']
-
-        for tag in new_tags:
-            assert tag not in preprint.tags.all().values_list('name', flat=True)
-
-        update_tags_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'tags': new_tags
-            }
-        )
-        res = app.patch_json_api(url, update_tags_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        preprint.reload()
-
-        assert sorted(
-            list(
-                preprint.tags.all().values_list(
-                    'name',
-                    flat=True))
-        ) == new_tags
-        assert mock_update_doi_metadata.called
-
-        # No tags
-        update_tags_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'tags': []
-            }
-        )
-        res = app.patch_json_api(url, update_tags_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        preprint.reload()
-        assert preprint.tags.count() == 0
-
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated')
-    def test_update_contributors(
-            self, mock_update_doi_metadata, app, user, preprint, url):
-        new_user = AuthUserFactory()
-        contributor_payload = {
-            'data': {
-                'attributes': {
-                    'bibliographic': True,
-                    'permission': WRITE,
-                    'send_email': False
-                },
-                'type': 'contributors',
-                'relationships': {
-                    'users': {
-                        'data': {
-                            'id': new_user._id,
-                            'type': 'users'
-                        }
-                    }
-                }
-            }
-        }
-
-        contributor_url = url + 'contributors/'
-
-        res = app.post_json_api(
-            contributor_url,
-            contributor_payload,
-            auth=user.auth)
-
-        assert res.status_code == 201
-        assert new_user in preprint.contributors
-        assert preprint.has_permission(new_user, WRITE)
-        assert PreprintContributor.objects.get(preprint=preprint, user=new_user).visible is True
-        assert mock_update_doi_metadata.called
-
-    def test_cannot_set_primary_file(self, app, user, preprint, url):
-        preprint.node = None
-        preprint.save()
-        #   test_write_contrib_can_attempt_to_set_primary_file
-        read_write_contrib = AuthUserFactory()
-        preprint.add_contributor(
-            read_write_contrib,
-            permissions=WRITE,
-            auth=Auth(user), save=True)
-        new_file = test_utils.create_test_preprint_file(
-            preprint, user, filename='lovechild_reason.pdf')
-
-        data = {
-            'data': {
-                'type': 'preprints',
-                'id': preprint._id,
-                'attributes': {},
-                'relationships': {
-                    'primary_file': {
-                        'data': {
-                            'type': 'file',
-                            'id': new_file._id
-                        }
-                    }
-                }
-            }
-        }
-
-        res = app.patch_json_api(
-            url, data,
-            auth=read_write_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 200
-
-    #   test_noncontrib_cannot_set_primary_file
-        non_contrib = AuthUserFactory()
-        new_file = test_utils.create_test_preprint_file(
-            preprint, user, filename='flowerchild_nik.pdf')
-
-        data = {
-            'data': {
-                'type': 'preprints',
-                'id': preprint._id,
-                'attributes': {},
-                'relationships': {
-                    'primary_file': {
-                        'data': {
-                            'type': 'file',
-                            'id': new_file._id
-                        }
-                    }
-                }
-            }
-        }
-
-        res = app.patch_json_api(
-            url, data,
-            auth=non_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-
-    def test_update_published(self, app, user):
-        unpublished = PreprintFactory(creator=user, is_published=False)
-        url = f'/{API_BASE}preprints/{unpublished._id}/'
-        payload = build_preprint_update_payload(
-            unpublished._id, attributes={'is_published': True})
-        app.patch_json_api(url, payload, auth=user.auth)
-        unpublished.reload()
-        assert unpublished.is_published
-
-    def test_update_published_does_not_make_node_public(
-            self, app, user):
-        project = ProjectFactory(creator=user)
-        unpublished = PreprintFactory(creator=user, is_published=False, project=project)
-        assert not unpublished.node.is_public
-        url = f'/{API_BASE}preprints/{unpublished._id}/'
-        payload = build_preprint_update_payload(
-            unpublished._id, attributes={'is_published': True})
-        app.patch_json_api(url, payload, auth=user.auth)
-        unpublished.node.reload()
-        unpublished.reload()
-
-        assert unpublished.node.is_public is False
-        assert unpublished.is_public
-
-    @responses.activate
-    @mock.patch('osf.models.preprint.update_or_enqueue_on_preprint_updated')
-    def test_update_preprint_task_called_on_api_update(
-            self, mock_on_preprint_updated, app, user, preprint, url):
-        responses.add(
-            responses.Response(
-                responses.POST,
-                CROSSREF_URL,
-                content_type='text/html;charset=ISO-8859-1',
-                status=200,
-            ),
-        )
-        update_doi_payload = build_preprint_update_payload(
-            preprint._id, attributes={'doi': '10.1235/ASDFASDF'})
-
-        app.patch_json_api(url, update_doi_payload, auth=user.auth)
-
-        assert mock_on_preprint_updated.called
-
-    def test_update_has_coi(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'has_coi': True}
-        )
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_coi']
-
-        preprint.reload()
-        assert preprint.has_coi
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_HAS_COI
-        assert log.params == {'preprint': preprint._id, 'user': user._id, 'value': True}
-
-    def test_update_conflict_of_interest_statement(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'conflict_of_interest_statement': 'Owns shares in Closed Science Corporation.'}
-        )
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        preprint.has_coi = False
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Cannot provide conflict of interest statement when has_coi is set to False.'
-
-        preprint.has_coi = True
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['conflict_of_interest_statement'] ==\
-               'Owns shares in Closed Science Corporation.'
-
-        preprint.reload()
-        assert preprint.conflict_of_interest_statement == 'Owns shares in Closed Science Corporation.'
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_COI_STATEMENT
-        assert log.params == {
-            'preprint': preprint._id,
-            'user': user._id,
-            'value': 'Owns shares in Closed Science Corporation.'
-        }
-
-    def test_update_has_data_links(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'has_data_links': 'available'})
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_data_links'] == 'available'
-
-        preprint.reload()
-        assert preprint.has_data_links
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_HAS_DATA_LINKS
-        assert log.params == {'value': 'available', 'user': user._id, 'preprint': preprint._id}
-
-    def test_update_why_no_data(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'why_no_data': 'My dog ate it.'})
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'You cannot edit this statement while your data links availability is set to true or is unanswered.'
-
-        preprint.has_data_links = 'no'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['why_no_data'] == 'My dog ate it.'
-
-        preprint.reload()
-        assert preprint.why_no_data
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_WHY_NO_DATA
-        assert log.params == {'user': user._id, 'preprint': preprint._id}
-
-    def test_update_data_links(self, app, user, preprint, url):
-        data_links = ['http://www.JasonKelce.com', 'http://www.ItsTheWholeTeam.com/']
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'data_links': data_links})
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        preprint.has_data_links = 'no'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Cannot provide data links when has_data_links is set to "no".'
-
-        preprint.has_data_links = 'available'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['data_links'] == data_links
-
-        preprint.reload()
-        assert preprint.data_links == data_links
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_DATA_LINKS
-        assert log.params == {'user': user._id, 'preprint': preprint._id}
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'data_links': 'maformed payload'})
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Expected a list of items but got type "str".'
-
-    def test_invalid_data_links(self, app, user, preprint, url):
-        preprint.has_data_links = 'available'
-        preprint.save()
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'data_links': ['thisaintright']})
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Enter a valid URL.'
-
-    def test_update_has_prereg_links(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'has_prereg_links': 'available'})
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_prereg_links'] == 'available'
-
-        preprint.reload()
-        assert preprint.has_prereg_links
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_HAS_PREREG_LINKS
-        assert log.params == {'value': 'available', 'user': user._id, 'preprint': preprint._id}
-
-    def test_invalid_prereg_links(self, app, user, preprint, url):
-        preprint.has_prereg_links = 'available'
-        preprint.save()
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'prereg_links': ['thisaintright']})
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Enter a valid URL.'
-
-    def test_no_data_links_clears_links(self, app, user, preprint, url):
-        preprint.has_data_links = 'available'
-        preprint.data_links = ['http://www.apple.com']
-        preprint.save()
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'has_data_links': 'no'})
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_data_links'] == 'no'
-        assert res.json['data']['attributes']['data_links'] == []
-
-        preprint.reload()
-        assert preprint.has_data_links == 'no'
-        assert preprint.data_links == []
-
-    def test_no_prereg_links_clears_links(self, app, user, preprint, url):
-        preprint.has_prereg_links = 'available'
-        preprint.prereg_links = ['http://example.com']
-        preprint.prereg_link_info = 'prereg_analysis'
-        preprint.save()
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'has_prereg_links': 'no'})
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_prereg_links'] == 'no'
-        preprint.reload()
-        assert res.json['data']['attributes']['prereg_links'] == []
-        assert not res.json['data']['attributes']['prereg_link_info']
-
-        assert preprint.prereg_links == []
-        assert preprint.prereg_link_info is None
-
-    def test_update_why_no_prereg(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'why_no_prereg': 'My dog ate it.'})
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        preprint.has_prereg_links = 'available'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'You cannot edit this statement while your prereg links availability is set to true or is unanswered.'
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={
-            'why_no_prereg': 'My dog ate it.',
-            'has_prereg_links': 'no'
-        })
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['why_no_prereg'] == 'My dog ate it.'
-
-        preprint.reload()
-        assert preprint.why_no_prereg == 'My dog ate it.'
-
-        log = preprint.logs.filter(action=PreprintLog.UPDATE_WHY_NO_PREREG).first()
-        assert log is not None, 'Expected log entry for why_no_prereg_updated not found.'
-        assert log.action == PreprintLog.UPDATE_WHY_NO_PREREG
-        assert log.params == {'user': user._id, 'preprint': preprint._id}
-
-    def test_update_prereg_links(self, app, user, preprint, url):
-
-        prereg_links = ['http://www.JasonKelce.com', 'http://www.ItsTheWholeTeam.com/']
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'prereg_links': prereg_links})
-
-        contrib = AuthUserFactory()
-        preprint.add_contributor(contrib, READ)
-        res = app.patch_json_api(url, update_payload, auth=contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'You do not have permission to perform this action.'
-
-        preprint.has_prereg_links = 'no'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'You cannot edit this field while your prereg links availability is set to false or is unanswered.'
-
-        preprint.has_prereg_links = 'available'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['prereg_links'] == prereg_links
-
-        preprint.reload()
-        assert preprint.prereg_links == prereg_links
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_PREREG_LINKS
-        assert log.params == {'user': user._id, 'preprint': preprint._id}
-
-        update_payload = build_preprint_update_payload(preprint._id, attributes={'prereg_links': 'maformed payload'})
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Expected a list of items but got type "str".'
-
-    def test_update_prereg_link_info(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'prereg_link_info': 'prereg_designs'}
-        )
-
-        preprint.has_prereg_links = 'no'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'You cannot edit this field while your prereg links availability is set to false or is unanswered.'
-
-        preprint.has_prereg_links = 'available'
-        preprint.save()
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['prereg_link_info'] == 'prereg_designs'
-
-        preprint.reload()
-        assert preprint.prereg_link_info == 'prereg_designs'
-        log = preprint.logs.first()
-        assert log.action == PreprintLog.UPDATE_PREREG_LINKS_INFO
-        assert log.params == {'user': user._id, 'preprint': preprint._id}
-
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'prereg_link_info': 'maformed payload'}
-        )
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == '"maformed payload" is not a valid choice.'
-
-    def test_update_has_coi_false_with_null_conflict_statement(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'has_coi': False,
-                'conflict_of_interest_statement': None
-            }
-        )
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_coi'] is False
-        assert res.json['data']['attributes']['conflict_of_interest_statement'] is None
-
-        preprint.reload()
-        assert preprint.has_coi is False
-        assert preprint.conflict_of_interest_statement is None
-
-    def test_update_has_data_links_no_with_data_links_provided(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'has_data_links': 'no',
-                'data_links': ['http://example.com/data']
-            }
-        )
-
-        initial_has_data_links = preprint.has_data_links
-        initial_data_links = preprint.data_links
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'Cannot provide data links when has_data_links is set to "no".'
-
-        preprint.reload()
-
-        assert preprint.has_data_links == initial_has_data_links
-        assert preprint.data_links == initial_data_links
-
-        assert preprint.has_data_links != 'no'
-
-    def test_update_has_data_links_no_with_empty_data_links(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'has_data_links': 'no',
-                'data_links': []
-            }
-        )
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_data_links'] == 'no'
-        assert res.json['data']['attributes']['data_links'] == []
-
-        preprint.reload()
-        assert preprint.has_data_links == 'no'
-        assert preprint.data_links == []
-
-    def test_update_has_prereg_links_no_with_empty_prereg_links(self, app, user, preprint, url):
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'has_prereg_links': 'no',
-                'prereg_links': [],
-                'prereg_link_info': ''
-            }
-        )
-
-        res = app.patch_json_api(url, update_payload, auth=user.auth)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_prereg_links'] == 'no'
-        assert res.json['data']['attributes']['prereg_links'] == []
-        assert res.json['data']['attributes']['prereg_link_info'] == ''
-
-        preprint.reload()
-        assert preprint.has_prereg_links == 'no'
-        assert preprint.prereg_links == []
-        assert preprint.prereg_link_info == ''
-
-    def test_non_admin_cannot_update_has_coi(self, app, user, preprint, url):
-        write_contrib = AuthUserFactory()
-        preprint.add_contributor(write_contrib, permissions=osf_permissions.WRITE, auth=Auth(user), save=True)
-
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={'has_coi': True}
-        )
-
-        res = app.patch_json_api(url, update_payload, auth=write_contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'Must have admin permissions to update author assertion fields.'
-
-        preprint.reload()
-        assert preprint.has_coi is None
-
-    def test_sloan_updates(self, app, user, preprint, url):
-        """
-        - Tests to ensure updating a preprint with unchanged data does not create superfluous log statements.
-        - Tests to ensure various dependent fields can be updated in a single request.
-        """
-        preprint.has_prereg_links = 'available'
-        preprint.prereg_links = ['http://no-sf.io']
-        preprint.prereg_link_info = 'prereg_designs'
-        preprint.save()
-
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'has_prereg_links': 'available',
-                'prereg_link_info': 'prereg_designs',
-                'prereg_links': ['http://osf.io'],  # changing here should be only non-factory created log.
-            }
-        )
-        app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        # Any superfluous log statements?
-        logs = preprint.logs.all().values_list('action', 'params')
-        assert logs.count() == 3  # actions should be: 'subjects_updated', 'published', 'prereg_links_updated'
-        assert logs.latest() == ('prereg_links_updated', {'user': user._id, 'preprint': preprint._id})
-
-        # Can we set `has_prereg_links` to false and update `why_no_prereg` in a single request?
-        update_payload = build_preprint_update_payload(
-            preprint._id,
-            attributes={
-                'has_prereg_links': 'no',
-                'why_no_prereg': 'My dog ate it.'
-            }
-        )
-        res = app.patch_json_api(url, update_payload, auth=user.auth, expect_errors=True)
-
-        assert res.status_code == 200
-        assert res.json['data']['attributes']['has_prereg_links'] == 'no'
-        assert res.json['data']['attributes']['why_no_prereg'] == 'My dog ate it.'
-
-        preprint.refresh_from_db()
-        assert preprint.has_prereg_links == 'no'
-        assert preprint.why_no_prereg == 'My dog ate it.'
-
-
-@pytest.mark.django_db
-class TestPreprintUpdateSubjects(UpdateSubjectsMixin):
-    @pytest.fixture()
-    def resource(self, user_admin_contrib, user_write_contrib, user_read_contrib):
-        preprint = PreprintFactory(creator=user_admin_contrib, is_published=True)
-        preprint.add_contributor(user_write_contrib, auth=Auth(user_admin_contrib))
-        preprint.add_contributor(
-            user_read_contrib,
-            auth=Auth(user_admin_contrib),
-            permissions=READ)
-        preprint.save()
-        return preprint
-
-
-@pytest.mark.django_db
-class TestPreprintUpdateLicense:
-
-    @pytest.fixture()
-    def admin_contrib(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def write_contrib(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def read_contrib(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def non_contrib(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def cc0_license(self):
-        return NodeLicense.objects.filter(name='CC0 1.0 Universal').first()
-
-    @pytest.fixture()
-    def mit_license(self):
-        return NodeLicense.objects.filter(name='MIT License').first()
-
-    @pytest.fixture()
-    def no_license(self):
-        return NodeLicense.objects.filter(name='No license').first()
-
-    @pytest.fixture()
-    def preprint_provider(self, cc0_license, no_license):
-        preprint_provider = PreprintProviderFactory()
-        preprint_provider.licenses_acceptable.add(*[cc0_license, no_license])
-        preprint_provider.save()
-        return preprint_provider
-
-    @pytest.fixture()
-    def preprint(
-            self, admin_contrib, write_contrib, read_contrib,
-            preprint_provider):
-        preprint = PreprintFactory(
-            creator=admin_contrib,
-            provider=preprint_provider)
-        preprint.add_contributor(write_contrib, permissions=WRITE, auth=Auth(admin_contrib))
-        preprint.add_contributor(
-            read_contrib,
-            auth=Auth(admin_contrib),
-            permissions=READ)
-        preprint.save()
-        return preprint
-
-    @pytest.fixture()
-    def url(self, preprint):
-        return f'/{API_BASE}preprints/{preprint._id}/'
-
-    @pytest.fixture()
-    def make_payload(self):
-        def payload(
-                node_id, license_id=None, license_year=None,
-                copyright_holders=None, jsonapi_type='preprints'
-        ):
-            attributes = {}
-
-            if license_year and copyright_holders:
-                attributes = {
-                    'license_record': {
-                        'year': license_year,
-                        'copyright_holders': copyright_holders
-                    }
-                }
-            elif license_year:
-                attributes = {
-                    'license_record': {
-                        'year': license_year
-                    }
-                }
-            elif copyright_holders:
-                attributes = {
-                    'license_record': {
-                        'copyright_holders': copyright_holders
-                    }
-                }
-
-            return {
-                'data': {
-                    'id': node_id,
-                    'type': jsonapi_type,
-                    'attributes': attributes,
-                    'relationships': {
-                        'license': {
-                            'data': {
-                                'type': 'licenses',
-                                'id': license_id
-                            }
-                        }
-                    }
-                }
-            } if license_id else {
-                'data': {
-                    'id': node_id,
-                    'type': jsonapi_type,
-                    'attributes': attributes
-                }
-            }
-
-        return payload
-
-    @pytest.fixture()
-    def make_request(self, app):
-        def request(url, data, auth=None, expect_errors=False):
-            return app.patch_json_api(
-                url, data, auth=auth, expect_errors=expect_errors)
-        return request
-
-    def test_admin_update_license_with_invalid_id(
-            self, admin_contrib, preprint, url, make_payload, make_request):
-        data = make_payload(
-            node_id=preprint._id,
-            license_id='thisisafakelicenseid'
-        )
-
-        assert preprint.license is None
-
-        res = make_request(
-            url, data,
-            auth=admin_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 404
-        assert res.json['errors'][0]['detail'] == 'Unable to find specified license.'
-
-        preprint.reload()
-        assert preprint.license is None
-
-    def test_admin_can_update_license(
-            self, admin_contrib, preprint, cc0_license,
-            url, make_payload, make_request):
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        assert preprint.license is None
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.reload()
-
-        res_data = res.json['data']
-        pp_license_id = preprint.license.node_license._id
-        assert res_data['relationships']['license']['data'].get(
-            'id', None) == pp_license_id
-        assert res_data['relationships']['license']['data'].get(
-            'type', None) == 'licenses'
-
-        assert preprint.license.node_license == cc0_license
-        assert preprint.license.year is None
-        assert preprint.license.copyright_holders == []
-
-        # check logs
-        log = preprint.logs.latest()
-        assert log.action == 'license_changed'
-        assert log.params.get('preprint') == preprint._id
-
-    def test_admin_can_update_license_record(
-            self, admin_contrib, preprint, no_license,
-            url, make_payload, make_request):
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=no_license._id,
-            license_year='2015',
-            copyright_holders=['Tonya Shepoly, Lucas Pucas']
-        )
-
-        assert preprint.license is None
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.reload()
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2015'
-        assert preprint.license.copyright_holders == [
-            'Tonya Shepoly, Lucas Pucas']
-
-    def test_cannot_update_license(
-            self, write_contrib, read_contrib, non_contrib,
-            preprint, cc0_license, url, make_payload, make_request):
-
-        #   test_write_contrib_can_update_license
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        res = make_request(
-            url, data,
-            auth=write_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 200
-        preprint.reload()
-
-        assert preprint.license.node_license == cc0_license
-
-    #   test_read_contrib_cannot_update_license
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        res = make_request(
-            url, data,
-            auth=read_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == exceptions.PermissionDenied.default_detail
-
-    #   test_non_contrib_cannot_update_license
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        res = make_request(
-            url, data,
-            auth=non_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == exceptions.PermissionDenied.default_detail
-
-    #   test_unauthenticated_user_cannot_update_license
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        res = make_request(url, data, expect_errors=True)
-        assert res.status_code == 401
-        assert res.json['errors'][0]['detail'] == exceptions.NotAuthenticated.default_detail
-
-    def test_update_error(
-            self, admin_contrib, preprint, preprint_provider,
-            mit_license, no_license, url, make_payload, make_request):
-
-        #   test_update_preprint_with_invalid_license_for_provider
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=mit_license._id
-        )
-
-        assert preprint.license is None
-
-        res = make_request(
-            url, data,
-            auth=admin_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-        assert res.json['errors'][0]['detail'] == 'Invalid license chosen for {}'.format(
-            preprint_provider.name)
-
-    #   test_update_preprint_license_without_required_year_in_payload
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=no_license._id,
-            copyright_holders=['Rachel', 'Rheisen']
-        )
-
-        res = make_request(
-            url, data,
-            auth=admin_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'year must be specified for this license'
-
-    #   test_update_preprint_license_without_required_copyright_holders_in_payload
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=no_license._id,
-            license_year='1994'
-        )
-
-        res = make_request(
-            url, data,
-            auth=admin_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 400
-        assert res.json['errors'][0]['detail'] == 'copyrightHolders must be specified for this license'
-
-    def test_update_preprint_with_existing_license_year_attribute_only(
-            self, admin_contrib, preprint, no_license, url, make_payload, make_request):
-        preprint.set_preprint_license(
-            {
-                'id': no_license.license_id,
-                'year': '2014',
-                'copyrightHolders': ['Daniel FromBrazil', 'Queen Jaedyn']
-            },
-            Auth(admin_contrib),
-        )
-        preprint.save()
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2014'
-        assert preprint.license.copyright_holders == [
-            'Daniel FromBrazil', 'Queen Jaedyn']
-
-        data = make_payload(
-            node_id=preprint._id,
-            license_year='2015'
-        )
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.license.reload()
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2015'
-        assert preprint.license.copyright_holders == [
-            'Daniel FromBrazil', 'Queen Jaedyn']
-
-    def test_update_preprint_with_existing_license_copyright_holders_attribute_only(
-            self, admin_contrib, preprint, no_license, url, make_payload, make_request):
-        preprint.set_preprint_license(
-            {
-                'id': no_license.license_id,
-                'year': '2014',
-                'copyrightHolders': ['Captain Haley', 'Keegor Cannoli']
-            },
-            Auth(admin_contrib),
-        )
-        preprint.save()
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2014'
-        assert preprint.license.copyright_holders == [
-            'Captain Haley', 'Keegor Cannoli']
-
-        data = make_payload(
-            node_id=preprint._id,
-            copyright_holders=['Reason Danish', 'Ben the NJB']
-        )
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.license.reload()
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2014'
-        assert preprint.license.copyright_holders == [
-            'Reason Danish', 'Ben the NJB']
-
-    def test_update_preprint_with_existing_license_relationship_only(
-            self, admin_contrib, preprint, cc0_license,
-            no_license, url, make_payload, make_request):
-        preprint.set_preprint_license(
-            {
-                'id': no_license.license_id,
-                'year': '2014',
-                'copyrightHolders': ['Reason', 'Mr. Lulu']
-            },
-            Auth(admin_contrib),
-        )
-        preprint.save()
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2014'
-        assert preprint.license.copyright_holders == ['Reason', 'Mr. Lulu']
-
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.license.reload()
-
-        assert preprint.license.node_license == cc0_license
-        assert preprint.license.year == '2014'
-        assert preprint.license.copyright_holders == ['Reason', 'Mr. Lulu']
-
-    def test_update_preprint_with_existing_license_relationship_and_attributes(
-            self, admin_contrib, preprint, cc0_license,
-            no_license, url, make_payload, make_request):
-        preprint.set_preprint_license(
-            {
-                'id': no_license.license_id,
-                'year': '2014',
-                'copyrightHolders': ['Reason', 'Mr. Cosgrove']
-            },
-            Auth(admin_contrib),
-            save=True
-        )
-
-        assert preprint.license.node_license == no_license
-        assert preprint.license.year == '2014'
-        assert preprint.license.copyright_holders == ['Reason', 'Mr. Cosgrove']
-
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id,
-            license_year='2015',
-            copyright_holders=['Rheisen', 'Princess Tyler']
-        )
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.license.reload()
-
-        assert preprint.license.node_license == cc0_license
-        assert preprint.license.year == '2015'
-        assert preprint.license.copyright_holders == [
-            'Rheisen', 'Princess Tyler']
-
-    def test_update_preprint_license_does_not_change_project_license(
-            self, admin_contrib, preprint, cc0_license,
-            no_license, url, make_payload, make_request):
-        project = ProjectFactory(creator=admin_contrib)
-        preprint.node = project
-        preprint.save()
-        preprint.node.set_node_license(
-            {
-                'id': no_license.license_id,
-                'year': '2015',
-                'copyrightHolders': ['Simba', 'Mufasa']
-            },
-            auth=Auth(admin_contrib)
-        )
-        preprint.node.save()
-        assert preprint.node.node_license.node_license == no_license
-
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=cc0_license._id
-        )
-
-        res = make_request(url, data, auth=admin_contrib.auth)
-        assert res.status_code == 200
-        preprint.reload()
-
-        assert preprint.license.node_license == cc0_license
-        assert preprint.node.node_license.node_license == no_license
-
-    def test_update_preprint_license_without_change_does_not_add_log(
-            self, admin_contrib, preprint, no_license, url, make_payload, make_request):
-        preprint.set_preprint_license(
-            {
-                'id': no_license.license_id,
-                'year': '2015',
-                'copyrightHolders': ['Kim', 'Kanye']
-            },
-            auth=Auth(admin_contrib),
-            save=True
-        )
-
-        before_num_logs = preprint.logs.count()
-        before_update_log = preprint.logs.latest()
-
-        data = make_payload(
-            node_id=preprint._id,
-            license_id=no_license._id,
-            license_year='2015',
-            copyright_holders=['Kanye', 'Kim']
-        )
-        res = make_request(url, data, auth=admin_contrib.auth)
-        preprint.reload()
-
-        after_num_logs = preprint.logs.count()
-        after_update_log = preprint.logs.latest()
-
-        assert res.status_code == 200
-        assert before_num_logs == after_num_logs
-        assert before_update_log._id == after_update_log._id
-
-
-@pytest.mark.django_db
 class TestPreprintDetailPermissions:
 
     @pytest.fixture()
@@ -2108,190 +480,3 @@ class TestPreprintDetailPermissions:
         unpublished.add_contributor(write_contrib, permissions=WRITE, save=True)
         res = app.get(url, auth=write_contrib.auth, expect_errors=True)
         assert res.status_code == 403
-
-
-@pytest.mark.django_db
-class TestReviewsPreprintDetailPermissions:
-
-    @pytest.fixture()
-    def admin(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def write_contrib(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def non_contrib(self):
-        return AuthUserFactory()
-
-    @pytest.fixture()
-    def public_project(self, admin):
-        return ProjectFactory(creator=admin, is_public=True)
-
-    @pytest.fixture()
-    def private_project(self, admin):
-        return ProjectFactory(creator=admin, is_public=False)
-
-    @pytest.fixture()
-    def subject(self):
-        return SubjectFactory()
-
-    @pytest.fixture()
-    def reviews_provider(self):
-        return PreprintProviderFactory(reviews_workflow='pre-moderation')
-
-    @pytest.fixture()
-    def unpublished_reviews_preprint(
-            self, admin, reviews_provider, subject, public_project, write_contrib):
-        preprint = PreprintFactory(
-            creator=admin,
-            filename='toe_socks_and_sunrises.pdf',
-            provider=reviews_provider,
-            subjects=[[subject._id]],
-            is_published=False,
-            machine_state=DefaultStates.PENDING.value)
-        preprint.add_contributor(write_contrib, permissions=WRITE)
-        preprint.save()
-        return preprint
-
-    @pytest.fixture()
-    def unpublished_reviews_initial_preprint(
-            self, admin, reviews_provider, subject, public_project):
-        return PreprintFactory(
-            creator=admin,
-            filename='toe_socks_and_sunrises.pdf',
-            provider=reviews_provider,
-            subjects=[[subject._id]],
-            is_published=False,
-            machine_state=DefaultStates.INITIAL.value)
-
-    @pytest.fixture()
-    def private_reviews_preprint(
-            self, admin, reviews_provider, subject, private_project, write_contrib):
-        preprint = PreprintFactory(
-            creator=admin,
-            filename='toe_socks_and_sunsets.pdf',
-            provider=reviews_provider,
-            subjects=[[subject._id]],
-            is_published=False,
-            is_public=False,
-            machine_state=DefaultStates.PENDING.value)
-        preprint.add_contributor(write_contrib, permissions=WRITE)
-        return preprint
-
-    @pytest.fixture()
-    def unpublished_url(self, unpublished_reviews_preprint):
-        return '/{}preprints/{}/'.format(
-            API_BASE, unpublished_reviews_preprint._id)
-
-    @pytest.fixture()
-    def unpublished_initial_url(self, unpublished_reviews_initial_preprint):
-        return '/{}preprints/{}/'.format(
-            API_BASE, unpublished_reviews_initial_preprint._id)
-
-    @pytest.fixture()
-    def private_url(self, private_reviews_preprint):
-        return '/{}preprints/{}/'.format(
-            API_BASE, private_reviews_preprint._id)
-
-    def test_reviews_preprint_is_published_detail(
-            self, app, admin, write_contrib, non_contrib,
-            unpublished_reviews_preprint, unpublished_url):
-
-        #   test_unpublished_visible_to_admins
-        res = app.get(unpublished_url, auth=admin.auth)
-        assert res.json['data']['id'] == unpublished_reviews_preprint._id
-
-    #   test_unpublished_visible_to_write_contribs
-        res = app.get(
-            unpublished_url,
-            auth=write_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 200
-
-    #   test_unpublished_invisible_to_non_contribs
-        res = app.get(
-            unpublished_url,
-            auth=non_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-
-    #   test_unpublished_invisible_to_public
-        res = app.get(unpublished_url, expect_errors=True)
-        assert res.status_code == 401
-
-    def test_reviews_preprint_initial_detail(
-            self, app, admin, write_contrib, non_contrib,
-            unpublished_reviews_initial_preprint,
-            unpublished_initial_url):
-
-        #   test_unpublished_visible_to_admins
-        res = app.get(unpublished_initial_url, auth=admin.auth)
-        assert res.json['data']['id'] == unpublished_reviews_initial_preprint._id
-
-    #   test_unpublished_invisible_to_write_contribs
-        res = app.get(
-            unpublished_initial_url,
-            auth=write_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-
-    #   test_unpublished_invisible_to_non_contribs
-        res = app.get(
-            unpublished_initial_url,
-            auth=non_contrib.auth,
-            expect_errors=True)
-        assert res.status_code == 403
-
-    #   test_unpublished_invisible_to_public
-        res = app.get(unpublished_initial_url, expect_errors=True)
-        assert res.status_code == 401
-
-    def test_reviews_preprint_is_public_detail(
-            self, app, admin, write_contrib, non_contrib,
-            private_reviews_preprint, private_url):
-
-        #   test_private_visible_to_admins
-        res = app.get(private_url, auth=admin.auth)
-        assert res.json['data']['id'] == private_reviews_preprint._id
-
-    #   test_private_visible_to_write_contribs
-        res = app.get(private_url, auth=write_contrib.auth, expect_errors=True)
-        assert res.status_code == 200
-
-    #   test_private_invisible_to_non_contribs
-        res = app.get(private_url, auth=non_contrib.auth, expect_errors=True)
-        assert res.status_code == 403
-
-    #   test_private_invisible_to_public
-        res = app.get(private_url, expect_errors=True)
-        assert res.status_code == 401
-
-
-@pytest.mark.django_db
-class TestPreprintDetailWithMetrics:
-    # enable the ELASTICSEARCH_METRICS switch for all tests
-    @pytest.fixture(autouse=True)
-    def enable_elasticsearch_metrics(self):
-        with override_switch(features.ELASTICSEARCH_METRICS, active=True):
-            yield
-
-    @pytest.mark.parametrize(('metric_name', 'metric_class_name'),
-    [
-        ('downloads', 'PreprintDownload'),
-        ('views', 'PreprintView'),
-    ])
-    def test_preprint_detail_with_downloads(self, app, settings, metric_name, metric_class_name):
-        preprint = PreprintFactory()
-        url = f'/{API_BASE}preprints/{preprint._id}/?metrics[{metric_name}]=total'
-
-        with mock.patch(f'api.preprints.views.{metric_class_name}.get_count_for_preprint') as mock_get_count_for_preprint:
-            mock_get_count_for_preprint.return_value = 42
-            res = app.get(url)
-
-        assert res.status_code == 200
-        data = res.json
-        assert 'metrics' in data['meta']
-        assert metric_name in data['meta']['metrics']
-        assert data['meta']['metrics'][metric_name] == 42
